feat!: support title as a rect element
Previously, we support title as a string or a list of string. This is
cubersome when writing complex layout, especially with lots of styling.
This change makes title more flexible as you can use any rect element as
the title.

BREAKING CHANGE: we don't support title as list anymore. Also,
attributes for title like color or text_modifiers have been removed in
favor of styling with rect element.